tried and tested new async converter and template

+++ UnitTests/project_files_test/Project_test.py
@@ -109,14 +109,14 @@ def create_mock_project_project_4():
         shutil.rmtree(project_path)
 
 @fixture
-def create_mock_project_eigen_referentie():
+async def create_mock_project_eigen_referentie():
     project_path = Path(PARENT_OF_THIS_FILE / 'OTLWizardProjects' / 'Projects' / 'eigen referentie')
     project_backup_path = Path(
         PARENT_OF_THIS_FILE / 'OTLWizardProjects' / 'Projects_backup' / project_path.name)
 
     if not project_path.exists():
         shutil.copytree(project_backup_path,project_path)
-        sleep(0.05) # sleep to give the system time to copy
+        await sleep(0.05) # sleep to give the system time to copy
     yield
 
     if project_path.exists():
@@ -598,7 +598,7 @@ def test_save_project_given_details(mock_project_home_path,create_mock_project_e
     shutil.rmtree(project_dir_path)
 
 @fixture
-def setup_quicksave_test_project(root_directory,mock_otl_wizard_dir) -> Project:
+async def setup_quicksave_test_project(root_directory,mock_otl_wizard_dir) -> Project:
     backup_quicksave_test_project_path  = Path(root_directory, "OTLWizardProjects", "Projects_backup", "quicksave_test")
     quicksave_test_project_path = Path(root_directory, "OTLWizardProjects", "Projects", "quicksave_test")
 
@@ -607,18 +607,17 @@ def setup_quicksave_test_project(root_directory,mock_otl_wizard_dir) -> Project:
 
     shutil.copytree(backup_quicksave_test_project_path,
                     quicksave_test_project_path)
-    sleep(0.5)  # sleep to give the system time to copy
-    setup_quicksave_test_project = Project.load_project(quicksave_test_project_path)
-
-    yield setup_quicksave_test_project
+    await sleep(0.5)  # sleep to give the system time to copy
+    yield Project.load_project(quicksave_test_project_path)
 
     if quicksave_test_project_path.exists():
         shutil.rmtree(quicksave_test_project_path)
 
-def test_load_validated_assets(setup_quicksave_test_project):
 
-
-    loaded_assets_and_relations:list[Union[RelatieObject, RelationInteractor]] = setup_quicksave_test_project.sync_load_validated_assets()
+@pytest.mark.asyncio
+async def test_load_validated_assets(setup_quicksave_test_project):
+    loaded_assets_and_relations:list[Union[RelatieObject, RelationInteractor]] = (
+        await setup_quicksave_test_project.load_validated_assets_async())
 
     print([str(asset) for asset in loaded_assets_and_relations])
 
@@ -930,9 +929,13 @@ def mock_get_last_quick_save_path_to_return_empty():
     yield
     Project.get_last_quick_save_path = original_get_last_quick_save_path
 
-def test_load_validated_assets_with_empty_path(setup_quicksave_test_project,mock_get_last_quick_save_path_to_return_empty):
 
-    loaded_assets_and_relations:list[Union[RelatieObject, RelationInteractor]] = setup_quicksave_test_project.sync_load_validated_assets()
+@pytest.mark.asyncio
+async def test_load_validated_assets_with_empty_path(setup_quicksave_test_project,
+                                                mock_get_last_quick_save_path_to_return_empty):
+
+    loaded_assets_and_relations:list[Union[RelatieObject, RelationInteractor]] = (
+        await setup_quicksave_test_project.load_validated_assets_async())
 
     assert loaded_assets_and_relations == []
 
@@ -952,7 +955,8 @@ def test_load_validated_assets_with_empty_path(setup_quicksave_test_project,mock
         "empty_project",
     ],indirect=True
 )
-def test_save_validated_assets_from_scratch(get_and_cleanup_empty_project: Project):
+@pytest.mark.asyncio
+async def test_save_validated_assets_from_scratch(get_and_cleanup_empty_project: Project):
     """
     testing to see if we can make the quicksave directory in an empty project
 
@@ -962,8 +966,7 @@ def test_save_validated_assets_from_scratch(get_and_cleanup_empty_project: Proje
     project, expected_res,param = get_and_cleanup_empty_project
     project.assets_in_memory = []
 
-    if project.quick_save_dir_path:
-        if project.quick_save_dir_path.exists():
+    if project.quick_save_dir_path and project.quick_save_dir_path.exists():
             shutil.rmtree(project.quick_save_dir_path)
 
     #to create the project folder we need to save it first
@@ -971,7 +974,7 @@ def test_save_validated_assets_from_scratch(get_and_cleanup_empty_project: Proje
 
     assert not project.quick_save_dir_path.exists()
 
-    project.save_validated_assets(asynchronous=False)
+    await project.save_validated_assets()
 
     # check that the quicksave directory exists now
     assert project.quick_save_dir_path.exists()
@@ -992,7 +995,8 @@ def test_save_validated_assets_from_scratch(get_and_cleanup_empty_project: Proje
         "empty_project",
     ],indirect=True
 )
-def test_save_validated_assets_delete_old_files(get_and_cleanup_empty_project: Project):
+@pytest.mark.asyncio
+async def test_save_validated_assets_delete_old_files(get_and_cleanup_empty_project: Project):
     """
     testing situations:
     1. when there is a file in the quicksave directory that doesn't follow the naming convention
@@ -1028,7 +1032,7 @@ def test_save_validated_assets_delete_old_files(get_and_cleanup_empty_project: P
     quick_save_files = os.listdir(project.get_quicksaves_dir_path())
     assert len(quick_save_files) == 3
 
-    project.save_validated_assets(asynchronous=False)
+    await project.save_validated_assets(asynchronous=False)
 
     # quicksave folder should contain
     # 1. strange_name_f
@@ -1041,23 +1045,25 @@ def test_save_validated_assets_delete_old_files(get_and_cleanup_empty_project: P
     expected_quick_save_files = sorted([f"quick_save-{young_date_str}.json", project.get_last_quick_save_path().name, 'strange_name.json'])
     assert quick_save_files == expected_quick_save_files
 @fixture
-def setup_preloaded_assets_in_memory(setup_quicksave_test_project) -> Project:
+async def setup_preloaded_assets_in_memory(setup_quicksave_test_project) -> Project:
 
-    setup_quicksave_test_project.assets_in_memory = setup_quicksave_test_project.sync_load_validated_assets()
+    setup_quicksave_test_project.assets_in_memory = await setup_quicksave_test_project.load_validated_assets_async()
     yield setup_quicksave_test_project
     setup_quicksave_test_project.assets_in_memory = []
 
-def test_save_validated_assets(setup_preloaded_assets_in_memory: Project):
+
+@pytest.mark.asyncio
+async def test_save_validated_assets(setup_preloaded_assets_in_memory: Project):
 
     expected_assets =  deepcopy(setup_preloaded_assets_in_memory.assets_in_memory)
     old_quicksave = deepcopy(setup_preloaded_assets_in_memory.last_quick_save)
 
-    setup_preloaded_assets_in_memory.save_validated_assets(asynchronous=False)
+    await setup_preloaded_assets_in_memory.save_validated_assets(asynchronous=False)
 
     # we expect there to be a new last_quick_save
     assert setup_preloaded_assets_in_memory.last_quick_save != old_quicksave
 
-    new_loaded_assets = setup_preloaded_assets_in_memory.sync_load_validated_assets()
+    new_loaded_assets = await setup_preloaded_assets_in_memory.load_validated_assets_async()
 
     assert new_loaded_assets == expected_assets
